# openapi/account-api.yaml
openapi: 3.1.0
info:
  title: Account Center Management API
  version: 1.0.0
  description: >-
    Account 管理端同源 Session API。浏览器不得持有应用 Secret、Session ID 或长期 Token。
    管理端 message/fieldErrors 根据可选 Accept-Language 返回中文或英文；支持 zh/en 地区变体与权重，
    缺失、不支持或非法头回退 zh-CN。响应包含 Content-Language 和 Vary: Accept-Language。
    错误码、字段名、状态值及用户业务数据不随语言变化。
servers:
  - url: /
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: 分页查询全局用户
      security:
        - sessionCookie: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: query
          in: query
          description: 匹配账号、姓名、邮箱或手机号
          schema:
            type: string
            maxLength: 128
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - createdAt,asc
              - createdAt,desc
              - account,asc
              - account,desc
              - name,asc
              - name,desc
            default: createdAt,desc
      responses:
        '200':
          description: 稳定分页结果；相同排序值始终以 id 升序作为次级排序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    post:
      tags: [Users]
      operationId: createUser
      summary: 创建全局用户
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 用户已创建；重复幂等请求复用首次响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /api/applications:
    get:
      tags: [Applications]
      operationId: listApplications
      summary: 查询已登记应用
      security:
        - sessionCookie: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
            maxLength: 128
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ApplicationStatus'
      responses:
        '200':
          description: 应用列表，任何响应均不包含 Secret
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    post:
      tags: [Applications]
      operationId: createApplication
      summary: 登记应用并生成首个 Secret
      description: >-
        Secret 仅在首次成功响应中显示。相同幂等键重放不会再次显示 Secret，返回 409；
        调用方遗失后必须执行轮换。
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: 应用已登记，Secret 只在本响应中显示
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationSecretResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/applications/{appCode}:
    get:
      tags: [Applications]
      operationId: getApplication
      summary: 查询应用详情
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/AppCode'
      responses:
        '200':
          description: 应用详情，不含 Secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Applications]
      operationId: updateApplication
      summary: 更新应用配置
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/AppCode'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApplicationRequest'
      responses:
        '200':
          description: 更新后的应用，version 单调递增
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/applications/{appCode}/status:
    put:
      tags: [Applications]
      operationId: changeApplicationStatus
      summary: 启用或禁用应用
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/AppCode'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeApplicationStatusRequest'
      responses:
        '200':
          description: 状态已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/applications/{appCode}/secret/rotate:
    post:
      tags: [Applications]
      operationId: rotateApplicationSecret
      summary: 轮换应用 Secret
      description: 新 Secret 仅显示一次，幂等重放不会再次返回敏感结果。
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/AppCode'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '200':
          description: Secret 已轮换
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationSecretResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/applications/{appCode}/secret/revoke:
    post:
      tags: [Applications]
      operationId: revokeApplicationSecret
      summary: 撤销应用 Secret
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/AppCode'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '200':
          description: Secret 已撤销，后续签名请求将被拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/users/{userId}/application-access:
    get:
      tags: [ApplicationAccess]
      operationId: listUserApplicationAccess
      summary: 查询用户对所有应用的准入期望状态
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: 未配置应用以 disabled、version=0 返回
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationAccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/{userId}/application-access/{appCode}:
    put:
      tags: [ApplicationAccess]
      operationId: changeUserApplicationAccess
      summary: 修改用户准入期望状态
      description: >-
        当前只持久化期望状态和待投递命令，不启动 Outbox Worker；
        integrationStatus 固定为 pending_application_adaptation。
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/AppCode'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeApplicationAccessRequest'
      responses:
        '202':
          description: 期望状态已保存，业务应用仍待适配
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationAccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
      description: 由 GET /api/session 初始化并只保存在前端内存中的同步器 Token
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
      description: 按操作者、方法和规范化资源路径隔离，完成结果至少保留 24 小时
    AppCode:
      name: appCode
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z][a-z0-9-]{0,63}$'
...
  responses:
    Unauthorized:
      description: 未登录或 Session 已失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: CSRF 校验失败或当前角色没有访问能力
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationFailed:
      description: 请求字段校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 账号重复、资源版本冲突或幂等键冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UserStatus:
      type: string
      enum: [enabled, disabled]
    UserResponse:
      type: object
      additionalProperties: false
      required: [id, account, email, name, phone, status, version]
      properties:
        id:
          type: string
          maxLength: 64
        account:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        version:
          type: integer
          format: int64
          minimum: 1
        createdAt:
          type: [string, 'null']
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [account, email, name, phone]
      properties:
        account:
          type: string
          minLength: 1
          maxLength: 128
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
        name:
          type: string
          minLength: 1
          maxLength: 128
        phone:
          type: string
          minLength: 1
          maxLength: 64
...
    UserPageResponse:
      type: object
      additionalProperties: false
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
    ApplicationStatus:
      type: string
      enum: [enabled, disabled]
...
    CreateApplicationRequest:
      type: object
      additionalProperties: false
      required:
        - appCode
        - name
        - entryUrl
        - ssoCallbackUrl
        - permissionIframeUrl
        - notifyBaseUrl
        - protocolCapabilities
      properties:
        appCode:
          type: string
          pattern: '^[a-z][a-z0-9-]{0,63}$'
        name:
          type: string
          minLength: 1
          maxLength: 128
        entryUrl:
          type: string
          format: uri
          maxLength: 512
        ssoCallbackUrl:
          type: string
          format: uri
          maxLength: 512
          description: 精确回调 URI，不接受通配符或 fragment
        permissionIframeUrl:
          type: string
          maxLength: 512
          description: 精确授权页 URI，可包含 {externalUserId} 路径占位符
        notifyBaseUrl:
          type: string
          format: uri
          maxLength: 512
        defaultTenantCode:
          type: [string, 'null']
          maxLength: 64
        protocolCapabilities:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ApplicationProtocolCapability'
    UpdateApplicationRequest:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/CreateApplicationFields'
        - type: object
          required: [version]
          properties:
            version:
              type: integer
              format: int64
              minimum: 1
...
    ApplicationResponse:
      type: object
      additionalProperties: false
      required:
        - appCode
        - name
        - entryUrl
        - ssoCallbackUrl
        - permissionIframeUrl
        - notifyBaseUrl
        - defaultTenantCode
        - status
        - version
        - secretVersion
        - secretState
        - protocolCapabilities
      properties:
        appCode:
          type: string
        name:
          type: string
        entryUrl:
          type: string
          format: uri
        ssoCallbackUrl:
          type: string
          format: uri
        permissionIframeUrl:
          type: string
        notifyBaseUrl:
          type: string
          format: uri
        defaultTenantCode:
          type: string
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        version:
          type: integer
          format: int64
          minimum: 1
        secretVersion:
          type: integer
          minimum: 1
        secretState:
          $ref: '#/components/schemas/ApplicationSecretState'
        protocolCapabilities:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ApplicationProtocolCapability'
        createdAt:
          type: [string, 'null']
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    ApplicationSecretResponse:
      type: object
      additionalProperties: false
      required: [application, secret]
      properties:
        application:
          $ref: '#/components/schemas/ApplicationResponse'
        secret:
          type: string
          minLength: 32
          description: 仅在本次成功响应中显示，不得写入浏览器持久化、日志或截图
    VersionedReasonRequest:
      type: object
      additionalProperties: false
      required: [version, reason]
      properties:
        version:
          type: integer
          format: int64
          minimum: 1
        reason:
          type: string
          minLength: 1
          maxLength: 256
...